requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.6
email-validator==2.1.0

# Database
supabase==2.3.4
postgrest==0.13.2

# AI/LLM
openai==2.14.0
google-generativeai==0.3.2  # For Gemini (optional)

# Agent Framework (will add LangGraph later after testing basic API)
# langgraph - to be added once we test basic functionality  
# langchain - to be added once we test basic functionality
# langchain-core - to be added once we test basic functionality

# Background Jobs
celery==5.3.6
redis==5.0.1

# Email
resend==0.7.0
# sendgrid==6.11.0  # Alternative email provider

# Date/Time Parsing
python-dateutil==2.8.2
dateparser==1.2.0

# HTTP Client
httpx<0.26,>=0.24

# Environment Variables
python-dotenv==1.0.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.0
structlog==24.1.0

# Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0  # Pre-built wheels available

# Rate Limiting
slowapi==0.1.9

# Utilities
pytz==2024.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
ruff==0.1.15
mypy==1.8.0
